cpukit/.../termios: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This is frequently an int or ssize_t variable compared to a uint32_t or
size_t. Sometimes the size_t is from a sizeof() use.

+++ cpukit/include/rtems/termiostypes.h
@@ -119,8 +119,8 @@ typedef struct rtems_termios_tty {
    * The canonical (cooked) character buffer
    */
   char    *cbuf;
-  int    ccount;
-  int    cindex;
+  size_t ccount;
+  size_t cindex;
 
   /*
    * Keep track of cursor (printhead) position
